requirements.txt
altgraph==0.17
pefile==2019.4.18
numpy==1.22.0
#opencv-python==4.2.0.32 - Could not find this older version. Using default (later) version 4.6.0.66
opencv-python
PyInstaller==3.6
pyserial==3.4
PySimpleGUI==4.18.2
python-chess==0.30.1
gTTS==2.3.1
pygame==2.1.3
